fix(macos): W3DDisplay, W3DGameEngine, windows.h stubs — iterative build fixes
- Guard IsIconic, CreateBMPFile in W3DDisplay.cpp with #ifndef __APPLE__
- Guard W3DWebBrowser include and CComObject in Win32GameEngine.h
- Add PBITMAPINFOHEADER/PBITMAPINFO/LPTR typedefs in windows.h
- Add cursor/window stubs: SetCursor, GetCursorPos, ScreenToClient, IsIconic
- Add message processing stubs: MSG, PeekMessage, TranslateMessage, DispatchMessage
- Add error mode, execution state, MessageBox constants
- Exclude Windows-only debug/profile source files from macOS build
- Add threading stubs: CreateEvent, SetEvent, WaitForSingleObject, _beginthread
- Fix pointer-to-int truncation in W3DAssetManager, SoundScene, W3DProgressBar
- Add windows.h PCH to z_wwaudio target
- Fix BINKEXPORT macro for non-Windows platforms
- Update FARPROC to function pointer type matching Windows behavior
- Generals g_gameengine: fix 40+ pointer casts, WOL guards, GameSpy exclusions

+++ Generals/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/SkirmishGameOptionsMenu.cpp
@@ -839,7 +839,7 @@ static void handlePlayerSelection(int index)
 	Int playerType, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
   UnicodeString title = GadgetComboBoxGetText(combo);
-	playerType = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	playerType = (Int)(intptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 	GameInfo *myGame = TheSkirmishGameInfo;
 
 	if (myGame)
@@ -858,7 +858,7 @@ static void handleColorSelection(int index)
 	GameWindow *combo = comboBoxColor[index];
 	Int color, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	color = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	color = (Int)(intptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 
 	GameInfo *myGame = TheSkirmishGameInfo;
 
@@ -898,7 +898,7 @@ static void handlePlayerTemplateSelection(int index)
 	GameWindow *combo = comboBoxPlayerTemplate[index];
 	Int playerTemplate, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	playerTemplate = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	playerTemplate = (Int)(intptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 	GameInfo *myGame = TheSkirmishGameInfo;
 
 	if (myGame)
@@ -951,7 +951,7 @@ static void handleTeamSelection(int index)
 	GameWindow *combo = comboBoxTeam[index];
 	Int team, selIndex;
 	GadgetComboBoxGetSelectedPos(combo, &selIndex);
-	team = (Int)GadgetComboBoxGetItemData(combo, selIndex);
+	team = (Int)(intptr_t)GadgetComboBoxGetItemData(combo, selIndex);
 	GameInfo *myGame = TheSkirmishGameInfo;
 
 	if (myGame)
